remove WalletSelector
