--- apps/web/src/components/Shared/Login/WalletSelector.tsx
@@ -1,120 +1,0 @@
-import { Spinner } from '@components/UI/Spinner';
-import { XCircleIcon } from '@heroicons/react/solid';
-import { Analytics } from '@lib/analytics';
-import onError from '@lib/onError';
-import { Trans } from '@lingui/macro';
-import { ERROR_MESSAGE } from 'data/constants';
-import { useAuthenticateMutation, useChallengeLazyQuery, useUserProfilesLazyQuery } from 'lens';
-import type { Dispatch, FC } from 'react';
-import { useEffect } from 'react';
-import toast from 'react-hot-toast';
-import { CHAIN_ID } from 'src/constants';
-import { useAppPersistStore, useAppStore } from 'src/store/app';
-import { useAuthStore } from 'src/store/auth';
-import { USER } from 'src/tracking';
-import { useAccount, useNetwork, useSignMessage, useSwitchNetwork } from 'wagmi';
-
-interface Props {
-  setHasProfile: Dispatch<boolean>;
-}
-
-const WalletSelector: FC<Props> = ({ setHasProfile }) => {
-  const setProfiles = useAppStore((state) => state.setProfiles);
-  const setCurrentProfile = useAppStore((state) => state.setCurrentProfile);
-  const setProfileId = useAppPersistStore((state) => state.setProfileId);
-  const setShowAuthModal = useAuthStore((state) => state.setShowAuthModal);
-
-  const { chain } = useNetwork();
-  const { address, connector, isConnected } = useAccount();
-  const { signMessageAsync } = useSignMessage({ onError });
-  const { switchNetwork } = useSwitchNetwork();
-  const [loadChallenge, { error: errorChallenge }] = useChallengeLazyQuery({
-    fetchPolicy: 'no-cache'
-  });
-  const [authenticate, { error: errorAuthenticate }] = useAuthenticateMutation();
-  const [getProfiles, { error: errorProfiles }] = useUserProfilesLazyQuery();
-
-  const handleSign = async () => {
-    let keepModal = false;
-    try {
-      // Get challenge
-      const challenge = await loadChallenge({
-        variables: { request: { address } }
-      });
-
-      if (!challenge?.data?.challenge?.text) {
-        return toast.error(ERROR_MESSAGE);
-      }
-
-      // Get signature
-      const signature = await signMessageAsync({
-        message: challenge?.data?.challenge?.text
-      });
-
-      // Auth user and set cookies
-      const auth = await authenticate({
-        variables: { request: { address, signature } }
-      });
-      localStorage.setItem('accessToken', auth.data?.authenticate.accessToken);
-      localStorage.setItem('refreshToken', auth.data?.authenticate.refreshToken);
-
-      // Get authed profiles
-      const { data: profilesData } = await getProfiles({
-        variables: { ownedBy: address }
-      });
-
-      if (profilesData?.profiles?.items?.length === 0) {
-        setHasProfile(false);
-        keepModal = true;
-      } else {
-        const profiles: any = profilesData?.profiles?.items
-          ?.slice()
-          ?.sort((a, b) => Number(a.id) - Number(b.id))
-          ?.sort((a, b) => (a.isDefault === b.isDefault ? 0 : a.isDefault ? -1 : 1));
-        const currentProfile = profiles[0];
-        setProfiles(profiles);
-        setCurrentProfile(currentProfile);
-        setProfileId(currentProfile.id);
-      }
-      Analytics.track(USER.SIWL);
-    } catch (error) {
-      console.error(error);
-    } finally {
-      if (!keepModal) {
-        setShowAuthModal(false);
-      }
-    }
-  };
-
-  useEffect(() => {
-    if (chain?.id === CHAIN_ID) {
-      if (connector?.id && isConnected) {
-        handleSign();
-      }
-    } else {
-      switchNetwork?.(CHAIN_ID);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isConnected, chain]);
-
-  return (
-    <div className="space-y-3">
-      <div className="space-y-2.5">
-        <div className="flex text-center text-sm font-bold">
-          <div className="flex">
-            <Spinner size="sm" className="mx-auto mr-2" />
-            <Trans>Signing …</Trans>
-          </div>
-        </div>
-      </div>
-      {(errorChallenge || errorAuthenticate || errorProfiles) && (
-        <div className="flex items-center space-x-1 font-bold text-red-500">
-          <XCircleIcon className="h-5 w-5" />
-          <div>{ERROR_MESSAGE}</div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default WalletSelector;
